add kinoData endpoints to client and renamed
